Safari-611.1.10/LayoutTests/accessibility/unknown-roles-not-exposed-expected.txt


This tests accessibility objects with role 'unknown' are not being exposed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS findUnknownChild(accessibilityController.focusedElement) == null is true
PASS successfullyParsed is true

TEST COMPLETE
